refactor: hash avatars with Bun.hash instead of hand-rolled FNV-1a
The avatar hash also ran in the browser (Avatar.svelte's fallback), so it
couldn't use Bun.hash — a Bun-only global. Make the server the sole
authority for avatar names: sanitizeInstance resolves each instance's
sprite name, so the client renders purely by name (findAvatar) and never
hashes. pickAvatar/pickUniqueAvatar move to a server-only module
(pick.server.ts) backed by Bun.hash.crc32; FNV-1a is deleted.

Per the chosen "full switch", instances created before this change (no
stored avatar) reshuffle once to their crc32 sprite; new instances
persist their name and are unaffected.
